requirements.txt
mysql-connector-python
pandas
ollama
python-dotenv 